cncli repo url change

+++ preprod/scripts/update_cncli.sh
@@ -1,16 +1,37 @@
 #!/bin/bash
 # Only relevant for block producing node
 
-# Installing CNCLI binary
+CNCLI_PATH=`which cncli`
+if [[ $CNCLI_PATH == "" ]]; then
+    echo "CNCLI does not seem to be installed. Please run init_cncli first."
+    exit 1
+fi
+
+CNCLI_VERSION=`cncli --version | head -1 | awk '{print $2}'`
 RELEASETAG=$(curl -s https://api.github.com/repos/cardano-community/cncli/releases/latest | jq -r .tag_name)
 VERSION=$(echo ${RELEASETAG} | cut -c 2-)
+
+if [[ $CNCLI_VERSION < $RELEASETAG ]]; then
+    echo "A newer version of CNCLI is available."
+
+    if ! promptyn "Please confirm you want to install version $RELEASETAG? (y/n)"; then
+        echo "Ok bye!"
+        exit 1
+    fi
+fi
+echo
+echo '---------------- Installing CNCLI binary ----------------'
+
+sudo systemctl stop cncli_sync.service
+
 echo "Installing CNCLI binary release ${RELEASETAG}"
-
 mkdir -p $HOME/download/cncli
-
 curl -sLJ https://github.com/cardano-community/cncli/releases/download/${RELEASETAG}/cncli-${VERSION}-x86_64-unknown-linux-gnu.tar.gz -o $HOME/download/cncli/cncli-${VERSION}-x86_64-unknown-linux-gnu.tar.gz
 
 sudo tar xzvf $HOME/download/cncli/cncli-${VERSION}-x86_64-unknown-linux-gnu.tar.gz -C /usr/local/bin/
 
 echo "Checking installed version:"
-cncli -V
+cncli --version
+
+sudo systemctl daemon-reload
+sudo systemctl start cncli_sync.service
